TestLogger is started
The proof of concept is done, not robust, but done. It could use some
work though.

Hasher and logger has been made where Hasher takes a path and returns a
dictionary og all the files in that path along with the md5 hash as a
value.
Logger takes the ram, cpu, harddisk, threads and handles every 5
seconds. The result  posted to a txt file